+++ combatris/src/game/panes/vote.h
@@ -9,7 +9,7 @@ class Vote {
  public:
   Vote() { Clear(); }
 
-  void Add(size_t host_id,CampaignType type, size_t seed) {
+  void Add(uint64_t host_id,CampaignType type, uint64_t seed) {
     auto& v = campaigns_.at(type);
 
     v.push_back(PlayerInfo(host_id, seed));
@@ -18,21 +18,21 @@ class Vote {
 
   void Clear();
 
-  std::optional<size_t> Cast(size_t host_id) const;
+  std::optional<uint64_t> Cast(uint64_t host_id) const;
 
  private:
   struct PlayerInfo {
-    PlayerInfo(size_t host_id, size_t seed) : host_id_(host_id), seed_(seed) {}
+    PlayerInfo(uint64_t host_id, uint64_t seed) : host_id_(host_id), seed_(seed) {}
 
     inline bool operator < (const PlayerInfo& pi) const { return pi.host_id_ < host_id_; }
 
-    inline bool operator == (size_t host_id) const { return host_id == host_id_; }
+    inline bool operator == (uint64_t host_id) const { return host_id == host_id_; }
 
-    size_t host_id_;
-    size_t seed_;
+    uint64_t host_id_;
+    uint64_t seed_;
   };
 
-  inline bool IsInList(size_t host_id, const std::vector<PlayerInfo>& v) const { return v.end() != std::find(v.begin(), v.end(), host_id); }
+  inline bool IsInList(uint64_t host_id, const std::vector<PlayerInfo>& v) const { return v.end() != std::find(v.begin(), v.end(), host_id); }
 
   std::map<CampaignType, std::vector<PlayerInfo>> campaigns_;
 };
